Refactor security analysis network loading
