+++ src/main/java/com/powsybl/openloadflow/network/impl/Networks.java
@@ -6,16 +6,12 @@
  */
 package com.powsybl.openloadflow.network.impl;
 
+import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.reporter.Reporter;
 import com.powsybl.iidm.network.*;
-import com.powsybl.openloadflow.network.LfNetwork;
-import com.powsybl.openloadflow.network.LfNetworkParameters;
-import com.powsybl.openloadflow.network.SlackBusSelector;
+import com.powsybl.openloadflow.network.*;
 
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -110,24 +106,65 @@ public final class Networks {
         return LfNetwork.load(network, new LfNetworkLoaderImpl(), parameters, reporter);
     }
 
+    private static void retainAndCloseNecessarySwitches(Network network, Set<Switch> switchesToOpen, Set<Switch> switchesToClose) {
+        network.getSwitchStream()
+                .filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
+                .forEach(sw -> sw.setRetained(false));
+        switchesToOpen.stream()
+                .filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
+                .forEach(sw -> sw.setRetained(true));
+        switchesToClose.stream()
+                .filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
+                .forEach(sw -> sw.setRetained(true));
+
+        switchesToClose.forEach(sw -> sw.setOpen(false)); // in order to be present in the network.
+    }
+
+    private static void restoreInitialTopology(LfNetwork network, Set<Switch> allSwitchesToClose) {
+        var connectivity = network.getConnectivity();
+        connectivity.startTemporaryChanges();
+        allSwitchesToClose.stream().map(Identifiable::getId).forEach(id -> {
+            LfBranch branch = network.getBranchById(id);
+            connectivity.removeEdge(branch);
+        });
+        Set<LfBus> removedBuses = connectivity.getVerticesRemovedFromMainComponent();
+        removedBuses.forEach(bus -> bus.setDisabled(true));
+        Set<LfBranch> removedBranches = new HashSet<>(connectivity.getEdgesRemovedFromMainComponent());
+        // we should manage branches open at one side.
+        for (LfBus bus : removedBuses) {
+            bus.getBranches().stream().filter(b -> !b.isConnectedAtBothSides()).forEach(removedBranches::add);
+        }
+        removedBranches.forEach(branch -> branch.setDisabled(true));
+    }
+
     public static LfNetworkList load(Network network, LfNetworkParameters networkParameters,
                                      Set<Switch> switchesToOpen, Set<Switch> switchesToClose, Reporter reporter) {
         if (switchesToOpen.isEmpty() && switchesToClose.isEmpty()) {
             return new LfNetworkList(load(network, networkParameters, reporter));
         } else {
+            if (!networkParameters.isBreakers()) {
+                throw new PowsyblException("LF networks have to be built from bus/breaker view");
+            }
+            // create a temporary working variant to build LF networks
             String tmpVariantId = "olf-tmp-" + UUID.randomUUID();
             String workingVariantId = network.getVariantManager().getWorkingVariantId();
             network.getVariantManager().cloneVariant(network.getVariantManager().getWorkingVariantId(), tmpVariantId);
             network.getVariantManager().setWorkingVariant(tmpVariantId);
-            network.getSwitchStream().filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
-                    .forEach(sw -> sw.setRetained(false));
-            switchesToOpen.stream().filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
-                    .forEach(sw -> sw.setRetained(true));
-            switchesToClose.stream().filter(sw -> sw.getVoltageLevel().getTopologyKind() == TopologyKind.NODE_BREAKER)
-                    .forEach(sw -> sw.setRetained(true));
-            switchesToClose.forEach(sw -> sw.setOpen(false)); // in order to be present in the network.
-            return new LfNetworkList(load(network, networkParameters, reporter),
-                                     new LfNetworkList.VariantCleaner(network, workingVariantId, tmpVariantId));
+
+            // retain in topology all switches that could be open or close
+            // and close switches that could be closed during the simulation
+            retainAndCloseNecessarySwitches(network, switchesToOpen, switchesToClose);
+
+            List<LfNetwork> lfNetworks = load(network, networkParameters, reporter);
+
+            if (!switchesToClose.isEmpty()) {
+                for (LfNetwork lfNetwork : lfNetworks) {
+                    // disable all buses and branches not connected to main component (because of switch to close)
+                    restoreInitialTopology(lfNetwork, switchesToClose);
+                }
+            }
+
+            return new LfNetworkList(lfNetworks, new LfNetworkList.VariantCleaner(network, workingVariantId, tmpVariantId));
         }
     }
 }
